Unbreak history handling of models with OneToOneFields
We add a new "costume" model that is a OneToOneField in Animal.  Because
we're dealing with cartoon animals, this makes sense :)

Because OneToOneField is a _subclass_ of ForeignKey, Binder assumed the
corresponding relation would have a many2many mapping back to the
original model.

This would cause it to instantiate an empty object of the new type, and
this would raise an error when referencing the reverse relation value.
This resulted in the surprising error message "Costume has no Animal",
when saving an animal, even if the animal has no costume!

Ref #25