Remove singleton pattern from Router class (#28).
Instead, we pass the router to the view class in kwargs, which it then
saves as a property on itself.

This allows us to instantiate multiple routers, for example in the
tests, but it also makes it easier to combine multiple apps which all
rely on Binder.

This is a backwards incompatible change!