Move filtering of fields into the filter classes.
The consequence of this change is that the filter class will no longer
get the field _object_ but the field _name_ and model.  This allows one
to instantiate a (or create a completely new custom) filter class, for
"virtual" fields on which we filter.

In short, this makes it really easy to create custom filters.