Add workaround + regression test for bug due to router change
Now that router is no longer a singleton, we need to set the router on
every view we construct.  If we don't, the view will crash whenever it
tries to call a method on its router.

A proper fix might entail refactoring router.model_view() to return an
instance of the view instead of the view class.  That way, the router
is responsible for passing itself to all views it constructs.