Properly fix the __ order_by issue (which had no ticket and no tests!)
In the previous "fix", parse_order_by accepted a sort of "mixed"
Binder and Django syntax that supports both dot and double underscore
notation for relations.  By doing so, it would now also be possible to
use double underscore in relations in the request's query string,
which we definitely don't want to expose.  This could lead to an
information leak if we tried to block certain field orderings in
views (the double underscore would bypass all Binder parsing).

This commit undoes 5065d38348 and 5065d38348 in favor of a simpler
solution: we unconditionally append the default ordering to the
user-supplied ordering to make the result set more stable.  We don't
even need to filter out any orderings that the user supplied, because
orders are applied from left to right, and if there's a total ordering
of a result set, any subsequent orderings are no-ops.

This commit also adds a regression test for the issue itself.