No longer allow custom date on history start + better date.
Binder history records the date/time on the Changeset. This used to be
the date/time of the history *start* (in history.start()).

However, we decided the date/time of the DB transaction commit would be
better. So now Binder history records the date at the end of
history.commit().

It also used to be possible to specify a custom date/time in
history.start(date=...); this has now been removed.