Do not materialize rel_id set when parsing "where"
This set may be quite huge!  For example where(deleted=False) might
return everything if nothing has been soft-deleted yet.  This list of
ids will completely defeat the query planner's optimization efforts.

So, instead, we keep the queryset as-is and let Django ORM generate a
subselect based on the referenced base model, which is much more
efficient in most typical cases.