Do not force evaluation of entire model queryset just before using values_list
When fetching the fk list of an m2m field, we only are interested in
the fk of the related model and the pk of the current model, not every
single field.  By forcing evaluation of the same queryset we use later
but with the added values_list(), we request the entire dataset
instead of just the two fields we are interested in.

This killed performance on a project where we have a large model with
the pk and fk of an m2m field in an index, which was defeated by
forcing every field to be fetched.

The forced queryset evaluation line was probably added for a reason,
but there are no tests breaking when dropping the line, so let's pray
and hope nothing breaks.  Otherwise, we'll find eventually figure the
reason this was added :)