Convert field validation errors to Binder Validation errors
This ensures our response has a 400 status code and a proper body with
error data when available.

Previously, we'd convert all validation errors *and* TypeErrors into a
simple RequestError, but that's not too helpful when filling out a
form and the user types in something invalid.  Usually we have JS
helpers which will format a field properly, always, but this may not
be the case for every single field type.

TypeErrors are (for now?) still treated as request errors.