Remove prefetch related and only select pk column when counting records
This seems to create slightly faster queries when they are complex, as
Django's ORM will wrap those in a subselect which does the count, and
that will be performing a count on a full record set while we're only
interested in the pk.  At best, the optimizer will see that it can do
an index scan instead of a full table scan.  At worst, it will
serialize a large list of integers in memory which is still much less
terrible than doing the same for a whole record set (which may need a
temporary table on disk).

This also adds a test suite for pagination, since we had none so far.

Maybe at some point we want to make the meta['total_records'] optional
so the frontend can decide when it is not interested in paginating.
For very large tables this can make quite the difference!