Add performance hack to avoid unnecessary serialize and re-fetch in _store
In a multi-put, the saved objects are not returned in the response.
However, _store will still re-fetch and serialize them via get_objs.

For large models with many fields or complex get_obj implementations
(or simply lots of saved objects), the re-fetch and especially the
serialization in get_objs can be a drag on performance.

Because adding yet another keyword arg to _store just for this would
be rather obnoxious (breaking lots of code), I've decided to store the
information of whether a request is a multi-PUT on the request object.
This is hacky, but still somewhat better than adding another keyword
argument.