Make "with" faster by fetching all ids at once and allow easier virtual withs
This avoids traversing the "with" list fully recursively for every
entry by first processing the withs so we know all the different
relations that are requested at every nesting level.

Then, we can fetch all the relations on the same prefix at once.

For example, if we have with=foo.bar,foo.qux the old code would
instantiate the foo view twice, and fetch all bar ids and then all qux
ids, in two queries.  Now, we do one aggregate query on foo to
retrieve the bars and quxes per main ID.  This is a big performance
boost if the query on the foo table is heavy, as we only need to hit
it twice instead of 1+N where N is the number of relations requested.

This also adds (partial) support for "virtual relations" as we've
implemented them in a handful of projects.  This is totally hacky and
probably a very bad idea, but it reduces the amount of code required
and halves the number of queries we have to do for virtual relations.

Unfortunately, this changes the internal API for get_withs(): it
changes the withs argument from a flat list of strings to a nested
dict of preprocessed withs.

It also slightly changes the exposed external API: For regular
relations, the id(s) are now also included in the data as well as in
the with.  This removes the necessity for the reverse related name
mapping feature, but let's keep it for now.  This should have no
effect on well-behaving clients, as the effective result should be the
same.  This also means that the resulting JSON will be slightly
bigger, but that should rarely be a problem.