Exclude nested relations from main object
Instead, just use the direct relations starting from this particular
object.  For instance, we don't want gate.keeper ending up in the zoo
endpoint's data results if we ask for with gate.keeper.  Instead, we
want gate to end up there and keeper in the gate.

Also add regression test.