Adjust tests for filterable relations to match new reality
Now, when you add a where() on a relation, even if the relation is a
straight foreign key, the corresponding key will be set to None when
the relation was with'ed and does not occur in the list.

This is slightly different from the original behaviour where you would
get the id but it would not be in the with map itself.