Only add distinct() when filtering on "many" relations that create JOINs
The distinct() call we had originally in get() will avoid duplicate
records in general and works fine.  However, it has a pretty heavy
performance impact when doing this for large tables.  So instead, we
add the distinct only when actually traversing a filter that generates
a JOIN that *may* produce duplicate records.