Apply several little hacks to make Binder semi-compatible with MySQL
Only "where" filtering of withs is not supported.

This uses GROUP_CONCAT instead of array_agg (which MySQL doesn't
have).  It also means that we can only support integer primary keys
(for now).

Permission test is skipped because it uses a model that relies on
Postgres-specific fields.  This will be changed in another commit.

Had to replace one TextField with a CharField because MySQL cannot
make a unique index on TEXT.  Limitations galore, but that's the
usual fare with MySQL.