Ensure we don't generate duplicate ids in our field mappings
These extra IDs are generated because we have JOINs on the main query
which produce duplicate records from _other_ fields.  Unfortunately,
array_agg doesn't accept a DISTINCT on other fields than the ORDER BY
which is kind of unexpected.  For more information, see this ML post:
https://www.postgresql.org/message-id/02af01cc6830$f671e500$e355af00$@yahoo.com

We did not catch this problem before, because the tests were written
without any particular ordering in mind, so we simply used set
comparisons.  In the end it's okay to have extra ids, but it does give
a lot of overhead, so we change the set comparisons to proper list
comparisons.