Refactor _filter_field and _filter_relation to return Q objects.
This avoids requiring a subquery in the with_ids array_agg call.
Without a subselect, things are a lot faster!

There are a few situations in virtual relations where this is not the
best way, but not many projects use virtual relations and most of
these cases can be worked around in Django 3 due to being able to
filter on boolean expressions anyway...