Fix stack overflow error with Loaded Values mixin and Binder History with only()
This is a pretty obscure issue, but easily triggered once you know
what causes it.  When using .only() in a query, the missing fields get
marked as "deferred".  When you try to access those fields, as
binder_concrete_fields_as_dict() would do, Django triggers a new
fetch_from_db(), but with only that one missing field.  This then
again triggers the from_db(), which will access the _other_ missing
fields, triggering from_db() just on those fields (on a new instance)
and so on and so forth, ad infinitum, until Python decides that enough
is enough and raises a maximum recursion depth error.

The solution for this is to skip fields that are deferred when
populating the dict, but only in the situations which might trigger
runaway recursion.  Then, when the field is requested (like in
get_old_values() or in the history updater), we need to handle the
situation that the field may be missing and either use the current
value (assuming no change was made; we simply don't know), refetch it,
or ignore it.  This needs to be determined on a case by case basis,
as each situation is subtly different.