Reduce number of queries performed by MultiPUT when locking objects
Instead of doing one query per object, do one query per model to lock
all the objects.  This should improve performance of multi-PUT a bit.

This is a refactor, not a semantic change, so this also adds a note
regarding an inconsistency in how view permissions are not checked
during multi-PUT, while they are checked during singular PUT/POST.