Fix loaded values mixin for migrations
When you use the in-memory fake representation of models during
migrations, the loaded values mixin is going to break due to the
binder_concrete_fields_as_dict method not being available.

Now, one is not likely to even need this in the first place, so it's
fine to just skip the loaded values setting.