Detect annoying situation when distinct() isn't called
In some cases, multiple records are returned but we don't call
distinct() because of performance reasons.  These situations are
triggered when someone puts a Q() filter on a one->many relation
instead of returning a queryset.  The entire reason Q() filters were
added was to improve performance, but it can only be done when the
filters are one->one.