Avoid validating foreign key checks if the key never changed
This leverages our custom full_clean implementation (#141) in
combination with the LoadedValues mixin.  This allows us to know when
a value did not change and avoid validating the foreign key just
before save, because the foreign key was fine as it was stored, so it
cannot suddenly become invalid.

For complex models with many relations or which are saved a lot, this
makes a big difference, as the reduced number of queries can make
things quite a bit more efficient.