Get rid of workaround for complex sorting and replace it with a proper implementation
When fetching relations through "with", we used to skip complex
Meta.ordering expressions because they couldn't be prefixed with the
relation's name.

This implementation calls "deconstruct" from the deconstructible
decorator which is used on all serializable values.  Meta.ordering
must be serializable by definition, because it needs to usable in
migrations, so we can safely rely on that for the order by stuff.