Add support for selective override of "needs clean validation check"
Sometimes, a field requires "clean" validation even if it didn't
change itself.  This happens when the field stores a relation and
there's a database constraint which depends on other fields as well,
like in a partially unique constraint.

Then, when the other fields change, it means the relation needs to be
cleaned anyway.