Django 5 update (#252)
* Django 5 update

- Update github ci testing matrix
Note that python 3.8 and django 5 are not compatible so add an exception
not to run that combination

- Update timezone import
Starting from django 5 the utc timezone should be imported from the
default python datetime library instead of the django specific library

- Fix get_fixed_timezone
The get_fixed_timezone method however is still needed from the django
library so import that seperatly

* Include python 3.7 and django 3 testing

This particular case is still used in older projects. So lets make sure
we test this combination as well.

* Testsuite for deb 9, 10, 11

Debian 9 (python 3.7) needs different OS version, does not run on latest anymore.

---------

Co-authored-by: Bob Booij-Liewes <bob@codeyellow.nl>
Co-authored-by: Pieter van Beerendonk <pieter@codeyellow.nl>